backend/requirements.txt
# supabase
python-dotenv
psycopg2